- aula 09102021 refatorarmetodo excluir

+++ algafood-api/src/main/java/com/algaworks/algafood/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,9 +34,15 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 		return em.merge(cozinha);
 	}
 
+	@Transactional
 	@Override
-	public void remover(Cozinha cozinha) {
-		buscar(cozinha.getId());
+	public void remover(Long id) {
+		Cozinha cozinha = buscar(id);
+		
+		if (cozinha == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		
 		em.remove(cozinha);
 		
 	}
